requirements.txt
asgiref==3.7.2
chardet==3.0.4
Django==5.0.1
django-cpf==0.1.0
django-localflavor==4.0
django-phonenumber-field==7.3.0
django-postalcodes==0.4.0
idna==2.7
localflavor==1.9
phonenumberslite==8.13.29
pillow==10.2.0
python-stdnum==1.19
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
urllib3==1.24.3
